Removed check for displaying book data based on availability
The books that wouldn't have been available in the library wouldn't have
been displayed in the home screen which might lead a user to believe
that a particular book is not present in the library which might not be
true.

Instead, this kind of a check will be implemented by using filters.